+++ app/models/__init__.py
@@ -1,6 +1,5 @@
 from .db import db
 from .user import User
 from .categorites import Category
-from .decks import Deck
+from .decks import Deck, DeckList, Added_Decks
 from .cards import Card
-from .decklists import DeckList
